prj: Refactor project structure Overhaul `ExplorerBrowser`, and add second Browser pane
