+++ src/electrifier/App.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using electrifier.Activation;
 using electrifier.Contracts.Services;
+using electrifier.Controls.Vanara.Services;
 using electrifier.Helpers;
 using electrifier.Models;
 using electrifier.Notifications;
@@ -10,7 +11,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.UI.Xaml;
 using System.Diagnostics;
-using electrifier.Controls.Vanara.Services;
 using WinUIEx;
 using static Microsoft.Extensions.Hosting.Host;
 using UnhandledExceptionEventArgs = Microsoft.UI.Xaml.UnhandledExceptionEventArgs;
@@ -42,7 +42,9 @@ public partial class App : Application
     public static ShellNamespaceService NamespaceService
     {
         get;
-    } = new() { };
+    } = new()
+    {
+    };
 
     public App()
     {
