+++ src/electrifier/Controls/Vanara/ExplorerBrowser.xaml.cs
@@ -137,7 +137,7 @@ public class BrowserItem(Shell32.PIDL pidl, bool isFolder)
     public readonly Shell32.PIDL PIDL = new(pidl);
     public string DisplayName => ShellItem.GetDisplayName(ShellItemDisplayString.NormalDisplay) ?? ShellItem.ToString();
     public ShellItem ShellItem = new(pidl);
-    public SoftwareBitmapSource SoftwareBitmapSource = isFolder ? ShellNamespaceService.DefaultFolderImageBitmapSource : ShellNamespaceService.DefaultDocumentAssocImageBitmapSource;
+    public SoftwareBitmapSource SoftwareBitmapSource = isFolder ? ShellNamespaceService.FolderBitmapSource : ShellNamespaceService.DocumentBitmapSource;
     public new ObservableCollection<BrowserItem> ChildItems = [];
     public static BrowserItem FromPIDL(Shell32.PIDL pidl) => new(pidl, false);
     public static BrowserItem FromShellFolder(ShellFolder shellFolder) => new(shellFolder.PIDL, true);
