+++ src/electrifier/Controls/Vanara/Services/IShellNamespaceService.cs
@@ -1,0 +1,3 @@
+﻿namespace electrifier.Controls.Vanara.Services;
+
+public interface IShellNamespaceService;
