+++ src/electrifier/Controls/Vanara/Services/NavigationLogEventArgs.cs
@@ -1,0 +1,23 @@
+﻿namespace electrifier.Controls.Vanara.Services;
+
+/// <summary>The event argument for NavigationLogChangedEvent</summary>
+public class NavigationLogEventArgs : EventArgs
+{
+    /// <summary>Indicates CanNavigateBackward has changed</summary>
+    public bool CanNavigateBackwardChanged
+    {
+        get; set;
+    }
+
+    /// <summary>Indicates CanNavigateForward has changed</summary>
+    public bool CanNavigateForwardChanged
+    {
+        get; set;
+    }
+
+    /// <summary>Indicates the Locations collection has changed</summary>
+    public bool LocationsChanged
+    {
+        get; set;
+    }
+}
