+++ src/electrifier/Controls/Vanara/ShellListView.xaml
@@ -5,7 +5,8 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:local="using:electrifier.Controls.Vanara"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             d:DataContext="{d:DesignInstance Type=local:ExplorerBrowser}"
+             xmlns:services="using:electrifier.Controls.Vanara.Services"
+             d:DataContext="{d:DesignInstance Type=services:ExplorerBrowser}"
              mc:Ignorable="d">
     <Grid>
         <Grid.Resources>
